[frontend] replace connector buttons into a menu (#7328)
